chore: refactor form fields
