+++ packages/amplify_authenticator/lib/amplify_authenticator.dart
@@ -393,7 +393,7 @@ class _AuthenticatorState extends State<Authenticator> {
               child: InheritedForms(
                 confirmSignInNewPasswordForm:
                     widget.confirmSignInNewPasswordForm,
-                resetPasswordForm: const ResetPasswordForm(),
+                resetPasswordForm: ResetPasswordForm(),
                 sendCodeForm: SendCodeForm(),
                 signInForm: widget.signInForm,
                 signUpForm: widget.signUpForm,
