+++ packages/amplify_authenticator/lib/src/enums/enums.dart
@@ -14,12 +14,8 @@
  */
 
 export 'auth_screen.dart';
+export 'authenticator_form_field_type.dart';
 export 'button_size.dart';
-export 'confirm_signin_types.dart';
-export 'confirm_signup_types.dart';
 export 'gender.dart';
-export 'reset_password_field.dart';
-export 'signin_types.dart';
-export 'signup_types.dart';
 export 'status_type.dart';
 export 'verify_attribute_field_types.dart';
